% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_genescan.R
\name{format_genescan}
\alias{format_genescan}
\title{This function takes a genescan file outputted from ThermoFishers Peakscanner application
as input. It assumes the user has converted the .csv to a .xlsx, but otherwise has not touched the file.
Required column names are 'Sample Filename', 'Dye Color', 'Size', and 'Height'. A period may be used
in place of whitespace. Outputs a .xlsx file containing one row for each sample, with columns for
the height and size of the top three peaks above a user-specified Size cutoff for a user specified Dye Colour.
Also contains a few extra, empty columns for downstream analysis our lab commonly uses
Recommend running with full file paths}
\usage{
format_genescan(file, out, sizeCutoff = 200, dye = "BLUE")
}
\arguments{
\item{file}{Path to the input file}

\item{out}{Path for the output to be written to}

\item{sizeCutoff}{Numeric value below which peaks will be filtered from the dataset. Default is 200}

\item{dye}{Character value to select which colour dye will be used to extract peaks. Default is 'BLUE'}
}
\value{
A .xslx file containing the top 3 peaks for each sample and their ratio. One row per sample
}
\description{
This function takes a genescan file outputted from ThermoFishers Peakscanner application
as input. It assumes the user has converted the .csv to a .xlsx, but otherwise has not touched the file.
Required column names are 'Sample Filename', 'Dye Color', 'Size', and 'Height'. A period may be used
in place of whitespace. Outputs a .xlsx file containing one row for each sample, with columns for
the height and size of the top three peaks above a user-specified Size cutoff for a user specified Dye Colour.
Also contains a few extra, empty columns for downstream analysis our lab commonly uses
Recommend running with full file paths
}
